+++ src/Recyclarr.Core/Config/Models/IServiceConfiguration.cs
@@ -1,9 +1,11 @@
+using System.IO.Abstractions;
 using Recyclarr.TrashGuide;
 
 namespace Recyclarr.Config.Models;
 
 public interface IServiceConfiguration
 {
+    IFileInfo? YamlPath { get; }
     SupportedServices ServiceType { get; }
     string InstanceName { get; }
     Uri BaseUrl { get; }
@@ -11,7 +13,6 @@ public interface IServiceConfiguration
     bool DeleteOldCustomFormats { get; }
     ICollection<CustomFormatConfig> CustomFormats { get; }
     QualityDefinitionConfig? QualityDefinition { get; }
-
     IReadOnlyCollection<QualityProfileConfig> QualityProfiles { get; }
     bool ReplaceExistingCustomFormats { get; }
 }