+++ src/Recyclarr.Core/CoreAutofacModule.cs
@@ -10,6 +10,7 @@ using Recyclarr.Compatibility;
 using Recyclarr.Compatibility.Radarr;
 using Recyclarr.Compatibility.Sonarr;
 using Recyclarr.Config;
+using Recyclarr.Config.Filtering;
 using Recyclarr.Config.Parsing;
 using Recyclarr.Config.Parsing.PostProcessing;
 using Recyclarr.Config.Parsing.PostProcessing.ConfigMerging;
@@ -99,13 +100,25 @@ public class CoreAutofacModule : Module
 
         builder.RegisterType<SecretsProvider>().As<ISecretsProvider>().SingleInstance();
         builder.RegisterType<YamlIncludeResolver>().As<IYamlIncludeResolver>();
-        builder.RegisterType<ConfigurationRegistry>().As<IConfigurationRegistry>();
-        builder.RegisterType<ConfigurationLoader>().As<IConfigurationLoader>();
+        builder.RegisterType<ConfigurationRegistry>();
+        builder.RegisterType<ConfigurationLoader>();
         builder.RegisterType<ConfigurationFinder>().As<IConfigurationFinder>();
         builder.RegisterType<ConfigValidationExecutor>();
         builder.RegisterType<ConfigParser>();
         builder.RegisterType<ConfigSaver>();
         builder.RegisterType<ConfigurationScopeFactory>();
+
+        // Filter Processors
+        builder.RegisterType<ConfigFilterProcessor>();
+        builder
+            .RegisterTypes(
+                typeof(NonExistentInstancesFilter),
+                typeof(DuplicateInstancesFilter),
+                typeof(SplitInstancesFilter),
+                typeof(InvalidInstancesFilter)
+            )
+            .As<IConfigFilter>()
+            .OrderByRegistration();
 
         // Keyed include processors
         builder
@@ -136,11 +149,12 @@ public class CoreAutofacModule : Module
             .As<IConfigDeprecationCheck>()
             .OrderByRegistration();
 
-        builder.RegisterType<RootConfigYamlValidator>().As<IValidator>();
-
         // These validators are required by IncludePostProcessor
         builder.RegisterType<RadarrConfigYamlValidator>().As<IValidator>();
         builder.RegisterType<SonarrConfigYamlValidator>().As<IValidator>();
+
+        // Required by ConfigurationRegistry
+        builder.RegisterType<ServiceConfigYamlValidator>().As<IValidator<ServiceConfigYaml>>();
     }
 
     private static void RegisterHttp(ContainerBuilder builder)