+++ tests/Recyclarr.IntegrationTests/IncludePostProcessorIntegrationTest.cs
@@ -16,7 +16,7 @@ public class IncludePostProcessorIntegrationTest : IntegrationTestFixture
 
         var config = new RootConfigYaml
         {
-            Radarr = new Dictionary<string, RadarrConfigYaml>
+            Radarr = new Dictionary<string, RadarrConfigYaml?>
             {
                 ["service1"] = new() { ApiKey = "asdf", BaseUrl = "fdsa" },
             },
@@ -45,7 +45,7 @@ public class IncludePostProcessorIntegrationTest : IntegrationTestFixture
 
         var config = new RootConfigYaml
         {
-            Radarr = new Dictionary<string, RadarrConfigYaml>
+            Radarr = new Dictionary<string, RadarrConfigYaml?>
             {
                 ["service1"] = new()
                 {
@@ -79,7 +79,7 @@ public class IncludePostProcessorIntegrationTest : IntegrationTestFixture
 
         var config = new RootConfigYaml
         {
-            Radarr = new Dictionary<string, RadarrConfigYaml>
+            Radarr = new Dictionary<string, RadarrConfigYaml?>
             {
                 ["service1"] = new()
                 {
@@ -134,7 +134,7 @@ public class IncludePostProcessorIntegrationTest : IntegrationTestFixture
 
         var config = new RootConfigYaml
         {
-            Radarr = new Dictionary<string, RadarrConfigYaml>
+            Radarr = new Dictionary<string, RadarrConfigYaml?>
             {
                 ["service1"] = new()
                 {
@@ -164,7 +164,7 @@ public class IncludePostProcessorIntegrationTest : IntegrationTestFixture
             .BeEquivalentTo(
                 new RootConfigYaml
                 {
-                    Radarr = new Dictionary<string, RadarrConfigYaml>
+                    Radarr = new Dictionary<string, RadarrConfigYaml?>
                     {
                         ["service1"] = new()
                         {