+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Improved error handling in YAML configuration as well as how those errors are rendered to console
+  output. (#396)
+
 ## [7.4.0] - 2024-11-11
 
 ### Added